# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Application API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the generated application",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "https://your-production-url.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/income-statement/{id}": {
      "get": {
        "summary": "Get income statement by ID",
        "description": "Retrieve a specific income statement record",
        "tags": [
          "Income Statement"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Income statement ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved income statement"
          },
          "404": {
            "description": "Income statement not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update income statement",
        "description": "Update an existing income statement record",
        "tags": [
          "Income Statement"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated"
          },
          "404": {
            "description": "Income statement not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete income statement",
        "description": "Delete an income statement record",
        "tags": [
          "Income Statement"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted"
          },
          "404": {
            "description": "Income statement not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/income-statement": {
      "get": {
        "summary": "Get all income statements",
        "description": "Retrieve all income statement records with pagination",
        "tags": [
          "Income Statement"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 100
            },
            "description": "Maximum number of records to return"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Number of records to skip"
          },
          {
            "in": "query",
            "name": "year",
            "schema": {
              "type": "string"
            },
            "description": "Filter by year"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved income statements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new income statement",
        "description": "Create a new income statement record",
        "tags": [
          "Income Statement"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "year",
                  "account_category"
                ],
                "properties": {
                  "year": {
                    "type": "string"
                  },
                  "account_category": {
                    "type": "string"
                  },
                  "sales_revenue": {
                    "type": "number"
                  },
                  "service_revenue": {
                    "type": "number"
                  },
                  "interest_revenue": {
                    "type": "number"
                  },
                  "other_revenue": {
                    "type": "number"
                  },
                  "total_revenues": {
                    "type": "number"
                  },
                  "advertising": {
                    "type": "number"
                  },
                  "bad_debt": {
                    "type": "number"
                  },
                  "commissions": {
                    "type": "number"
                  },
                  "cost_of_goods_sold": {
                    "type": "number"
                  },
                  "depreciation": {
                    "type": "number"
                  },
                  "employee_benefits": {
                    "type": "number"
                  },
                  "furniture_and_equipment": {
                    "type": "number"
                  },
                  "insurance": {
                    "type": "number"
                  },
                  "interest_expense": {
                    "type": "number"
                  },
                  "maintenance_and_repairs": {
                    "type": "number"
                  },
                  "office_supplies": {
                    "type": "number"
                  },
                  "payroll_taxes": {
                    "type": "number"
                  },
                  "rent": {
                    "type": "number"
                  },
                  "research_and_development": {
                    "type": "number"
                  },
                  "salaries_and_wages": {
                    "type": "number"
                  },
                  "software": {
                    "type": "number"
                  },
                  "travel": {
                    "type": "number"
                  },
                  "utilities": {
                    "type": "number"
                  },
                  "web_hosting_and_domains": {
                    "type": "number"
                  },
                  "other_expenses": {
                    "type": "number"
                  },
                  "total_expenses": {
                    "type": "number"
                  },
                  "net_income_before_taxes": {
                    "type": "number"
                  },
                  "income_tax_expense": {
                    "type": "number"
                  },
                  "income_from_continuing_operations": {
                    "type": "number"
                  },
                  "net_income": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Income statement created successfully"
          },
          "400": {
            "description": "Bad request - validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
...
}